+++ src/main/java/dev/pcvolkmer/mv64e/datamapper/mapper/MtbEpisodeDataMapper.java
@@ -20,7 +20,6 @@
 
 package dev.pcvolkmer.mv64e.datamapper.mapper;
 
-import dev.pcvolkmer.mv64e.datamapper.PropertyCatalogue;
 import dev.pcvolkmer.mv64e.datamapper.datacatalogues.KpaCatalogue;
 import dev.pcvolkmer.mv64e.mtb.MtbEpisodeOfCare;
 import dev.pcvolkmer.mv64e.mtb.PeriodDate;
@@ -36,12 +35,9 @@ import java.util.List;
 public class MtbEpisodeDataMapper implements DataMapper<MtbEpisodeOfCare> {
 
   private final KpaCatalogue kpaCatalogue;
-  private final PropertyCatalogue propertyCatalogue;
 
-  public MtbEpisodeDataMapper(
-      final KpaCatalogue kpaCatalogue, final PropertyCatalogue propertyCatalogue) {
+  public MtbEpisodeDataMapper(final KpaCatalogue kpaCatalogue) {
     this.kpaCatalogue = kpaCatalogue;
-    this.propertyCatalogue = propertyCatalogue;
   }
 
   /**