+++ src/test/java/dev/pcvolkmer/mv64e/datamapper/mapper/MtbEpisodeDataMapperTest.java
@@ -24,7 +24,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Mockito.doAnswer;
 
-import dev.pcvolkmer.mv64e.datamapper.PropertyCatalogue;
 import dev.pcvolkmer.mv64e.datamapper.datacatalogues.KpaCatalogue;
 import dev.pcvolkmer.mv64e.datamapper.test.Column;
 import dev.pcvolkmer.mv64e.datamapper.test.DateColumn;
@@ -44,15 +43,13 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class MtbEpisodeDataMapperTest {
 
   KpaCatalogue kpaCatalogue;
-  PropertyCatalogue propertyCatalogue;
 
   MtbEpisodeDataMapper dataMapper;
 
   @BeforeEach
-  void setUp(@Mock KpaCatalogue kpaCatalogue, @Mock PropertyCatalogue propertyCatalogue) {
+  void setUp(@Mock KpaCatalogue kpaCatalogue) {
     this.kpaCatalogue = kpaCatalogue;
-    this.propertyCatalogue = propertyCatalogue;
-    this.dataMapper = new MtbEpisodeDataMapper(kpaCatalogue, propertyCatalogue);
+    this.dataMapper = new MtbEpisodeDataMapper(kpaCatalogue);
   }
 
   @Test