+++ src/test/java/dev/pcvolkmer/mv64e/datamapper/test/Column.java
@@ -30,6 +30,7 @@ public class Column {
 
   public static final String ID = "id";
   public static final String HAUPTPROZEDUR_ID = "hauptprozedur_id";
+  public static final String PATIENT_ID = "patient_id";
   public static final String PATIENTEN_ID = "patienten_id";
 
   String name;